requirements.txt
google-genai>=0.2.0
chromadb>=0.4.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0


